+++ frontend/src/app/router/lazy/employer.pages.jsx
@@ -19,4 +19,5 @@ export const EmployerPasswordLoginPage = lazy(() => import('@/pages/employer/app
 export const EmployerCompanySettingsPage = lazy(() => import('@/pages/employer/app/account/CompanySettings'))
 export const EmployerBusinessLicensePage = lazy(() => import('@/pages/employer/app/account/BusinessLicense'))
 export const EmployerPersonalDataProtectionPage = lazy(() => import('@/pages/employer/app/account/PersonalDataProtection'))
+export const EmployerRecruitmentDemandPage = lazy(() => import('@/pages/employer/app/account/RecruitmentDemand'))
 export const EmployerGeneralSettingsPage = lazy(() => import('@/pages/employer/app/account/GeneralSetting'))